delete old article_delete_button.jsx
