+++ public/packs/manifest.json
@@ -1,6 +1,6 @@
 {
-  "application.js": "/packs/application-2b2f487d1c5a7dea8197.js",
-  "application.js.map": "/packs/application-2b2f487d1c5a7dea8197.js.map",
-  "server_rendering.js": "/packs/server_rendering-001cd3f56d3666b8f7dd.js",
-  "server_rendering.js.map": "/packs/server_rendering-001cd3f56d3666b8f7dd.js.map"
+  "application.js": "/packs/application-ae1fc68af64242b182ba.js",
+  "application.js.map": "/packs/application-ae1fc68af64242b182ba.js.map",
+  "server_rendering.js": "/packs/server_rendering-d031723b430c0d1e5391.js",
+  "server_rendering.js.map": "/packs/server_rendering-d031723b430c0d1e5391.js.map"
 }
